requirements.txt
Python 3.6.9
numpy 1.18.1
pytorch 1.5.1+cu101